+++ Dockerfile.data_cube_creation
@@ -17,7 +17,7 @@ RUN pip3 install -r requirements.txt
 ADD backend/__init__.py backend/__init__.py
 ADD backend/wmg backend/wmg
 ADD backend/corpus_asset_pipelines backend/corpus_asset_pipelines
-ADD backend/corpora backend/corpora
+ADD backend/common backend/common
 ADD backend/ontology_files backend/ontology_files
 
 ARG HAPPY_BRANCH="unknown"
