+++ Dockerfile.processing_image
@@ -22,8 +22,7 @@ RUN pip3 install -r requirements.txt
 
 ADD tests /single-cell-data-portal/tests
 ADD backend/__init__.py backend/__init__.py
-ADD backend/corpora/__init__.py backend/corpora/__init__.py
-ADD backend/corpora/dataset_processing backend/corpora/dataset_processing
+ADD backend/dataset_processing backend/dataset_processing
 ADD backend/common backend/common
 
 ARG HAPPY_BRANCH="unknown"
@@ -33,4 +32,4 @@ LABEL commit=${HAPPY_COMMIT}
 ENV COMMIT_SHA=${HAPPY_COMMIT}
 ENV COMMIT_BRANCH=${HAPPY_BRANCH}
 
-CMD ["python3", "-m", "backend.corpora.dataset_processing.process"]
+CMD ["python3", "-m", "backend.dataset_processing.process"]
