+++ Dockerfile.upload_failures
@@ -7,7 +7,7 @@ COPY backend/corpora/upload_failures .
 COPY backend/__init__.py ./backend/__init__.py
 COPY backend/corpora/__init__.py ./backend/corpora/__init__.py
 COPY backend/common ./backend/common
-COPY backend/corpora/dataset_processing ./backend/corpora/dataset_processing
+COPY backend/dataset_processing ./backend/dataset_processing
 
 ARG HAPPY_BRANCH="unknown"
 ARG HAPPY_COMMIT=""
