+++ scripts/cxg_remaster.py
@@ -6,7 +6,7 @@ import sys
 pkg_root = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))  # noqa
 sys.path.insert(0, pkg_root)  # noqa
 
-from backend.corpora.dataset_processing.remaster_cxg import compute
+from backend.dataset_processing.remaster_cxg import compute
 
 """
 Re-master an existing "X" tiledb array in a CXG
