+++ tests/unit/backend/corpora/dataset_processing/test_h5ad_data_file.py
@@ -10,7 +10,7 @@ from pandas import Series, DataFrame
 import tiledb
 
 from backend.common.utils.corpora_constants import CorporaConstants
-from backend.corpora.dataset_processing.h5ad_data_file import H5ADDataFile
+from backend.dataset_processing.h5ad_data_file import H5ADDataFile
 
 from tests.unit.backend.corpora.fixtures.environment_setup import fixture_file_path
