+++ tests/unit/processing_container/test_process.py
@@ -4,7 +4,7 @@ from unittest.mock import patch
 
 import requests
 
-from backend.corpora.dataset_processing import process
+from backend.dataset_processing import process
 from tests.unit.backend.fixtures.mock_aws_test_case import CorporaTestCaseUsingMockAWS
 
 
@@ -51,7 +51,7 @@ class TestDatasetProcessing(CorporaTestCaseUsingMockAWS):
             if os.path.exists(f):
                 os.remove(f)
 
-    @patch("backend.corpora.dataset_processing.process.download_from_source_uri")
+    @patch("backend.dataset_processing.process.download_from_source_uri")
     def test_main(self, mock_download_from_source_uri):
         """
         Tests full pipeline for processing an uploaded H5AD file, including database updates
